+++ src/game/gamerules/GameRuleFactory.ts
@@ -1,0 +1,5 @@
+import { GameRuleBase } from "./GameRuleBase";
+
+export abstract class GameRuleFactory {
+  abstract newGameRuleProxy(uuid: string): GameRuleBase
+}
