+++ src/game/gamerules/local/GuessNumber/bot_template/cpp/bot.h
