+++ src/game/gamerules/local/GuessNumber/bot_template/cpp/player_proxy_client.h
