+++ src/game/gamerules/local/GuessNumber/bots/cpp/CMakeLists.txt
