+++ src/game/gamerules/local/GuessNumber/bots/cpp/bot2/CMakeLists.txt
