+++ src/game/gamerules/local/GuessNumber/bots/cpp/common.cmake
