+++ src/game/gamerules/local/GuessNumber/bots/cpp/cppbot.cc
