+++ src/game/gamerules/local/GuessNumber/bots/ts/BinarySearchBotUseTemplate.test.ts
