+++ src/game/players/playerProxy/grpc/typescript/google/protobuf/struct.ts
